requirements.txt
# Core dependencies
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3

# Database
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.12
Flask-Migrate==4.0.4
alembic==1.10.4
psycopg2-binary==2.9.6

# Environment variables
python-dotenv==1.0.0

# Slack
slack-sdk==3.21.0
slack-bolt==1.18.0

# Production server
gunicorn==21.2.0

# Testing
pytest==7.3.1
pytest-flask==1.2.0